+++ Utilities/itktiff/tif_strip.c
@@ -38,7 +38,7 @@ summarize(TIFF* tif, size_t summand1, size_t summand2, const char* where)
    * XXX: We are using casting to uint32 here, bacause sizeof(size_t)
    * may be larger than sizeof(uint32) on 64-bit architectures.
    */
-  uint32  bytes = summand1 + summand2;
+  uint32  bytes = (uint32)(summand1) + (uint32)(summand2);
 
   if (bytes - summand1 != summand2) {
     TIFFError(tif->tif_name, "Integer overflow in %s", where);
@@ -51,7 +51,7 @@ summarize(TIFF* tif, size_t summand1, size_t summand2, const char* where)
 static uint32
 multiply(TIFF* tif, size_t nmemb, size_t elem_size, const char* where)
 {
-  uint32  bytes = nmemb * elem_size;
+  uint32  bytes = (uint32)(nmemb) * (uint32)(elem_size);
 
   if (elem_size && bytes / elem_size != nmemb) {
     TIFFError(tif->tif_name, "Integer overflow in %s", where);
