+++ Utilities/openjpeg/event.c
@@ -106,7 +106,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
     /* initialize the optional parameter list */
     va_start(arg, fmt);
     /* check the length of the format string */
-    str_length = (strlen(fmt) > MSG_SIZE) ? MSG_SIZE : strlen(fmt);
+    str_length = (int)((strlen(fmt) > MSG_SIZE) ? MSG_SIZE : strlen(fmt));
     /* parse the format string and put the result in 'message' */
     for (i = 0, j = 0; i < str_length; ++i) {
       if (fmt[i] == '%') {
@@ -120,7 +120,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
               char tmp[16];
               _itoa(va_arg(arg, int), tmp, 8);
               strcat(message, tmp);
-              j += strlen(tmp);
+              j += (int)strlen(tmp);
               ++i;
               break;
             }
@@ -130,7 +130,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
               char tmp[16];
               _itoa(va_arg(arg, int), tmp, 10);
               strcat(message, tmp);
-              j += strlen(tmp);
+              j += (int)strlen(tmp);
               ++i;
               break;
             }
@@ -139,7 +139,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
               char tmp[16];
               _itoa(va_arg(arg, int), tmp, 16);
               strcat(message, tmp);
-              j += strlen(tmp);
+              j += (int)strlen(tmp);
               ++i;
               break;
             }
@@ -147,7 +147,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
             {
               char *tmp = va_arg(arg, char*);
               strcat(message, tmp);
-              j += strlen(tmp);
+              j += (int)strlen(tmp);
               ++i;
               break;
             }
@@ -157,7 +157,7 @@ bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ...) {
               double value = va_arg(arg, double);
               sprintf(tmp, "%f", value);
               strcat(message, tmp);
-              j += strlen(tmp);
+              j += (int)strlen(tmp);
               ++i;
               break;
             }
