+++ Utilities/openjpeg/t2.c
@@ -288,7 +288,7 @@ static int t2_encode_packet(opj_tcd_tile_t * tile, opj_tcp_t * tcp, opj_pi_itera
     }
   }
   
-  return (c - dest);
+  return (int)(c - dest);
 }
 
 static void t2_init_seg(opj_tcd_seg_t * seg, int cblksty, int first) {
@@ -372,7 +372,7 @@ static int t2_decode_packet(opj_t2_t* t2, unsigned char *src, int len, opj_tcd_t
     bio_init_dec(bio, hd, tcp->ppt_len);
   } else {      /* Normal Case */
     hd = c;
-    bio_init_dec(bio, hd, src+len-hd);
+    bio_init_dec(bio, hd, (int)(src+len-hd));
   }
   
   present = bio_read(bio, 1);
@@ -393,17 +393,17 @@ static int t2_decode_packet(opj_t2_t* t2, unsigned char *src, int len, opj_tcd_t
     }
     
     if (cp->ppm == 1) {    /* PPM case */
-      cp->ppm_len += cp->ppm_data-hd;
+      cp->ppm_len += (int)(cp->ppm_data-hd);
       cp->ppm_data = hd;
-      return (c - src);
+      return (int)(c - src);
     }
     if (tcp->ppt == 1) {  /* PPT case */
-      tcp->ppt_len+=tcp->ppt_data-hd;
+      tcp->ppt_len += (int)(tcp->ppt_data-hd);
       tcp->ppt_data = hd;
-      return (c - src);
+      return (int)(c - src);
     }
     
-    return (hd - src);
+    return (int)(hd - src);
   }
   
   for (bandno = 0; bandno < res->numbands; bandno++) {
@@ -483,10 +483,10 @@ static int t2_decode_packet(opj_t2_t* t2, unsigned char *src, int len, opj_tcd_t
   }
   
   if (cp->ppm==1) {
-    cp->ppm_len+=cp->ppm_data-hd;
+    cp->ppm_len += (int)(cp->ppm_data-hd);
     cp->ppm_data = hd;
   } else if (tcp->ppt == 1) {
-    tcp->ppt_len+=tcp->ppt_data-hd;
+    tcp->ppt_len += (int)(tcp->ppt_data-hd);
     tcp->ppt_data = hd;
   } else {
     c=hd;
@@ -537,7 +537,7 @@ static int t2_decode_packet(opj_t2_t* t2, unsigned char *src, int len, opj_tcd_t
     }
   }
   
-  return (c - src);
+  return (int)(c - src);
 }
 
 /* ----------------------------------------------------------------------- */
@@ -565,7 +565,7 @@ int t2_encode_packets(opj_t2_t* t2, int tileno, opj_tcd_tile_t *tile, int maxlay
   for (pino = 0; pino <= cp->tcps[tileno].numpocs; pino++) {
     while (pi_next(&pi[pino])) {
       if (pi[pino].layno < maxlayers) {
-        e = t2_encode_packet(tile, &cp->tcps[tileno], &pi[pino], c, dest + len - c, image_info, tileno);
+        e = t2_encode_packet(tile, &cp->tcps[tileno], &pi[pino], c, (int)(dest + len - c), image_info, tileno);
         if (e == -999) {
           break;
         } else {
@@ -599,7 +599,7 @@ int t2_encode_packets(opj_t2_t* t2, int tileno, opj_tcd_tile_t *tile, int maxlay
     return e;
   }
 
-    return (c - dest);
+    return (int)(c - dest);
 }
 
 int t2_decode_packets(opj_t2_t *t2, unsigned char *src, int len, int tileno, opj_tcd_tile_t *tile) {
@@ -621,7 +621,7 @@ int t2_decode_packets(opj_t2_t *t2, unsigned char *src, int len, int tileno, opj
   for (pino = 0; pino <= cp->tcps[tileno].numpocs; pino++) {
     while (pi_next(&pi[pino])) {
       if ((cp->layer==0) || (cp->layer>=((pi[pino].layno)+1))) {
-        e = t2_decode_packet(t2, c, src + len - c, tile, &cp->tcps[tileno], &pi[pino]);
+        e = t2_decode_packet(t2, c, (int)(src + len - c), tile, &cp->tcps[tileno], &pi[pino]);
       } else {
         e = 0;
       }
@@ -648,7 +648,7 @@ int t2_decode_packets(opj_t2_t *t2, unsigned char *src, int len, int tileno, opj
     return e;
   }
   
-    return (c - src);
+    return (int)(c - src);
 }
 
 /* ----------------------------------------------------------------------- */
